code/jobs/launch_GPU_training/ressources/requirements.txt
huggingface_hub==0.14.1
evaluate==0.4.0
s3fs==2023.1.0
datasets==2.12.0
transformers==4.18.0
boto3==1.24.0
scikit-learn==1.2.2
saagieapi==0.9.1
mlflow-saagie==2.0.1